Fixed warning. JIRA: CXX-4973
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@63056 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objects/seq/seq_id_tree.hpp
@@ -736,7 +736,7 @@ public:
     const string& GetStrSuffix(void) const {
         return m_Key.m_StrSuffix;
     }
-    int GetStrDigits(void) const {
+    size_t GetStrDigits(void) const {
         return m_Key.GetStrDigits();
     }
     void Restore(CDbtag& id, TPacked param) const;
